src/auth/pyproject.toml
[project]
name = "auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "bcrypt<4.0.0",
    "fastapi[standard]>=0.136.1",
    "grpcio>=1.80.0",
    "grpcio-tools>=1.80.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.14.0",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy[asyncio]>=2.0.49",
    "uvicorn>=0.46.0",
]
